Move library to /vendor
As part of Treble project in Android O, all the device specific files have
to be located in a separate vendor partition. This is done by setting
LOCAL_PROPRIETARY_MODULE (the name is misleading). This change will not
break existing platforms without a vendor partition as it will just move
files to /system/vendor.

Change-Id: Ie051c479d300022bd8f35d2c4aaddee97c24009b
Signed-off-by: Rob Herring <robh@kernel.org>